+++ dsrnascan/__init__.py
@@ -1,5 +1,5 @@
 # dsRNAscan package
 from .dsRNAscan import main
 
-__version__ = '0.4.7'
+__version__ = '0.4.8'
 __all__ = ['main']